ivy_bt/requirements.txt
pandas
numpy
yfinance
matplotlib
seaborn
statsmodels
plotly
scikit-learn
pandas_ta
pydantic
pyyaml
pyarrow
